requirements_rocm.txt
# Core dependencies with ROCm support
torch-hip==2.5.0
torchvision-hip==0.20.0
torchaudio-hip==2.5.0

# Machine learning libraries
diffusers==0.31.0
transformers==4.46.1
accelerate==1.0.1
safetensors==0.4.5
huggingface-hub==0.26.2
sentencepiece==0.1.99
protobuf
grpcio
grpcio-tools

# Utility libraries (same as requirements.txt)
numpy==2.1.2
Pillow==11.0.0
psutil==6.1.0
tqdm==4.66.6
fsspec==2024.10.0
PyYAML==6.0.2
networkx==3.4.2
sympy==1.13.1
filelock==3.16.1
typing_extensions==4.12.2
regex==2024.9.11
nltk==3.9.1
jinja2==3.1.4

# Web and API libraries (same as requirements.txt)
gradio==5.4.0
fastapi==0.115.3
uvicorn==0.32.0
python-multipart==0.0.12
aiofiles==23.2.1
orjson==3.10.10
websockets==12.0

# Image processing libraries (same as requirements.txt)
opencv-python==4.8.1.78
scikit-image==0.21.0

# Miscellaneous libraries (same as requirements.txt)
cachetools==5.3.1
rich==13.9.3

# Development tools (same as requirements.txt)
ruff==0.7.1
black==24.10.0

# Testing tools (same as requirements.txt)
pytest==7.4.2
pytest-cov==6.0.0
mypy==1.13.0
